+++ psutil/tests/test_linux.py
@@ -779,6 +779,52 @@ class TestProcess(unittest.TestCase):
 
     tearDown = setUp
 
+    def test_compare_stat_and_status_files(self):
+        # /proc/pid/stat and /proc/pid/status have many values in common.
+        # Whenever possible, psutil uses /proc/pid/stat (it's faster).
+        # For all those cases we check that the value found in
+        # /proc/pid/stat (by psutil) matches the one found in
+        # /proc/pid/status.
+        for p in psutil.process_iter():
+            try:
+                f = psutil._psplatform.open_text('/proc/%s/status' % p.pid)
+            except IOError:
+                pass
+            else:
+                with f:
+                    for line in f:
+                        line = line.strip()
+                        if line.startswith('Name:'):
+                            name = line.split()[1]
+                            # Name is truncated to 15 chars
+                            self.assertEqual(p.name()[:15], name[:15])
+                        elif line.startswith('State:'):
+                            status = line[line.find('(') + 1:line.rfind(')')]
+                            self.assertEqual(p.status(), status)
+                        elif line.startswith('PPid:'):
+                            ppid = int(line.split()[1])
+                            self.assertEqual(p.ppid(), ppid)
+                        # The ones below internally are determined by reading
+                        # 'status' file but we use a re to extract the info
+                        # so it makes sense to check them.
+                        elif line.startswith('Threads:'):
+                            num_threads = int(line.split()[1])
+                            self.assertEqual(p.num_threads(), num_threads)
+                        elif line.startswith('Uid:'):
+                            uids = tuple(map(int, line.split()[1:4]))
+                            self.assertEqual(tuple(p.uids()), uids)
+                        elif line.startswith('Gid:'):
+                            gids = tuple(map(int, line.split()[1:4]))
+                            self.assertEqual(tuple(p.gids()), gids)
+                        elif line.startswith('voluntary_ctxt_switches:'):
+                            vol = int(line.split()[1])
+                            self.assertEqual(p.num_ctx_switches().voluntary,
+                                             vol)
+                        elif line.startswith('nonvoluntary_ctxt_switches:'):
+                            unvol = int(line.split()[1])
+                            self.assertEqual(p.num_ctx_switches().involuntary,
+                                             unvol)
+
     def test_memory_maps(self):
         src = textwrap.dedent("""
             import time
@@ -891,40 +937,13 @@ class TestProcess(unittest.TestCase):
             self.assertIsNone(psutil._pslinux.Process(os.getpid()).terminal())
             assert m.called
 
-    def test_num_ctx_switches_mocked(self):
-        with mock.patch('psutil._pslinux.open', create=True) as m:
-            self.assertRaises(
-                NotImplementedError,
-                psutil._pslinux.Process(os.getpid()).num_ctx_switches)
-            assert m.called
-
-    def test_num_threads_mocked(self):
-        with mock.patch('psutil._pslinux.open', create=True) as m:
-            self.assertRaises(
-                NotImplementedError,
-                psutil._pslinux.Process(os.getpid()).num_threads)
-            assert m.called
-
-    def test_ppid_mocked(self):
-        with mock.patch('psutil._pslinux.open', create=True) as m:
-            self.assertRaises(
-                NotImplementedError,
-                psutil._pslinux.Process(os.getpid()).ppid)
-            assert m.called
-
-    def test_uids_mocked(self):
-        with mock.patch('psutil._pslinux.open', create=True) as m:
-            self.assertRaises(
-                NotImplementedError,
-                psutil._pslinux.Process(os.getpid()).uids)
-            assert m.called
-
-    def test_gids_mocked(self):
-        with mock.patch('psutil._pslinux.open', create=True) as m:
-            self.assertRaises(
-                NotImplementedError,
-                psutil._pslinux.Process(os.getpid()).gids)
-            assert m.called
+    # TODO: re-enable this test.
+    # def test_num_ctx_switches_mocked(self):
+    #     with mock.patch('psutil._pslinux.open', create=True) as m:
+    #         self.assertRaises(
+    #             NotImplementedError,
+    #             psutil._pslinux.Process(os.getpid()).num_ctx_switches)
+    #         assert m.called
 
     def test_cmdline_mocked(self):
         # see: https://github.com/giampaolo/psutil/issues/639