+++ HISTORY.rst
@@ -8,6 +8,11 @@ Bug tracker at https://github.com/giampaolo/psutil/issues
 - #795: [Windows] new APIs to deal with Windows services: win_service_iter()
   and win_service_get().
 - #800: [Linux] psutil.virtual_memory() returns a new "shared" memory field.
+- XXX: [Linux] speedup /proc parsing:
+  - Process.ppid() is 20% faster
+  - Process.status() is 28% faster
+  - Process.name() is 25% faster
+  - Process.num_threads is 20% faster on Python 3
 
 **Bug fixes**